+++ server/pipeline_match.go
@@ -209,6 +209,20 @@ func (p *Pipeline) matchJoin(logger *zap.Logger, session Session, envelope *rtap
 			return
 		}
 		meta = &m
+	} else if mode == StreamModeMatchAuthoritative {
+		// The user was already in the match, and it's an authoritative match.
+		// Look up the match label to return it anyway.
+		l, err := p.matchRegistry.GetMatchLabel(session.Context(), matchID, node)
+		if err != nil {
+			// There was a problem looking up the label.
+			logger.Error("Error looking up match label", zap.String("match_id", matchIDString), zap.String("node", node), zap.Error(err))
+			session.Send(false, 0, &rtapi.Envelope{Cid: envelope.Cid, Message: &rtapi.Envelope_Error{Error: &rtapi.Error{
+				Code:    int32(rtapi.Error_RUNTIME_EXCEPTION),
+				Message: "Match label lookup failed.",
+			}}})
+			return
+		}
+		label = &wrappers.StringValue{Value: l}
 	}
 
 	// Whether the user has just (successfully) joined the match or was already a member, return the match info anyway.